Initial implementation of OverridableCanister
This PR creates a new class, `OverridableCanister`,
that allows values to be temporarily overridden for
testing purposes, with automatic rollback if you
use the block syntax with `#override`.

**Note** that it might be better to rename `Canister` to
`SingleCanister` or something, and then have what
is currently `OverridableCanister` just be `Canister`.

No matter what, the name `OverridableCanister` kinda stinks.
Ideas welcome.

Should be a drop-in replacement for a Canister with
the option to override values temporarily for
testing or whatnot.

First pass includes all the functionality and docs,
along with running all the existing specs against both
Canister and OverridableCanister.